refactor: simplify canvas page after vote extraction
