+++ frontend/src/features/gameplay/vote/index.ts
@@ -4,6 +4,7 @@ export { default as ColorPalette } from "./components/ColorPalette";
 export { default as VotePanel } from "./components/VotePanel";
 export { default as VotePopup } from "./components/VotePopup";
 
+export { default as useVotePopup } from "./hooks/useVotePopup";
 export { default as useVoteState } from "./hooks/useVoteState";
 export { default as useVoteTickets } from "./hooks/useVoteTickets";
